+++ braintree/transaction_review.py
@@ -1,0 +1,9 @@
+from braintree.resource import Resource
+
+class TransactionReview(Resource):
+    """
+    A class representing a Transaction Review.
+    """
+
+    def __init__(self, attributes):
+        Resource.__init__(self, None, attributes)